Fix: bundle usage

+++ tests_ruby/test_install_json/install_json.sh
@@ -3,5 +3,11 @@
 apt-get update
 apt-get install -y libgmp-dev
 rm -rf vendor
-bundle install --path=vendor
-bundle exec ruby -e "require 'json'; puts ({hello: 'world'}).to_json"
+# check ruby version
+if [ "$(ruby -e "Gem::Version.new('$VERSION') >= Gem::Version.new('2.4') ? puts('0') : puts('1')")" -ne 0 ]
+then
+	bundle install --path=vendor
+else
+	bundle config set path vendor
+	bundle install
+fi
